pyproject.toml
[project]
name = "invest-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.9",
    "langchain>=1.2.17",
    "langchain-anthropic>=1.4.1",
    "langchain-chroma>=1.1.0",
    "langchain-core>=1.3.2",
    "langchain-google-genai>=4.2.2",
    "langchain-huggingface>=1.2.2",
    "langchain-text-splitters>=1.1.2",
    "langgraph>=1.1.9",
    "langsmith>=0.7.37",
    "prompt-toolkit>=3.0.52",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "yfinance>=0.2.0",
]
